Feature(calendar-widget): Update the NDMI layer as well
This one is a bit hacky, and I'm sure could be done in a number of different ways.

I chose, in this commit, to hard code the possible dates for NDMI data, query them on each date selection, and update the url and layers param accordingly. However, this doesn't appear to work, as I seem to get back .exe files instead of .png files for all usdm layers except for "current".

That said, this is how I think it should work, and it's worth checking in for review.